feat: update system instructions, move home page
