+++ src/hooks.ts
@@ -1,0 +1,8 @@
+import { useEffect } from "react";
+import { refresh } from "./actions/refresh";
+
+export function useRefreshAiContext() {
+    useEffect(() => {
+        refresh().then();
+    }, []);
+}
